requirements.txt
numpy 
tensorflow==2.10.0
Flask==2.2.2
keras==2.10.0
Keras-Preprocessing==1.1.2
Werkzeug==2.2.2
Pillow 
